[Refactor] BasicInfoSetUpView 코드 분리 및 폴더 이동
